+++ lib/typings/src/public_api.ts
@@ -24,3 +24,4 @@ export { Any } from './types/any';
 export { PlainObjectOf } from './types/plaing-object-of';
 export { ClassType } from './types/class';
 export { MappedState } from './types/mapped-state';
+export { NgxsDataExtension } from './types/extension';